remove dfr_core.py
